% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_from_lm}
\alias{cache_from_lm}
\title{Create a sensitivity cache from an \code{lm} model}
\usage{
cache_from_lm(model, treatment)
}
\arguments{
\item{model}{An unweighted \code{\link[stats:lm]{stats::lm()}} object.}

\item{treatment}{Name of the single treatment model-matrix column.}
}
\value{
An \code{rv_cache} with the original model metadata attached.
}
\description{
Create a sensitivity cache from an \code{lm} model
}
