% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{point_estimate_rv}
\alias{point_estimate_rv}
\title{Point-estimate robustness value}
\usage{
point_estimate_rv(cache, null = 0)
}
\arguments{
\item{cache}{An \code{rv_cache}.}

\item{null}{Target coefficient value.}
}
\description{
Point-estimate robustness value
}
