% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.robust_rv_result}
\alias{print.robust_rv_result}
\title{Print a robustness-value result}
\usage{
\method{print}{robust_rv_result}(x, ...)
}
\arguments{
\item{x}{A \code{robust_rv_result}.}

\item{...}{Unused.}
}
\description{
Print a robustness-value result
}
