% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tidy_rv}
\alias{tidy_rv}
\title{Tidy one robustness-value result}
\usage{
tidy_rv(x)
}
\arguments{
\item{x}{A \code{robust_rv_result}.}
}
\value{
A one-row data frame.
}
\description{
Tidy one robustness-value result
}
